+++ batch3dfier/bag3d.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 
-"""Generate a 3D BAG data set"""
+"""Generate a 3D BAG data set. Linux only."""
 
 import os.path
-from io import StringIO
-from subprocess import call
+from subprocess import run
 
 from psycopg2 import sql
 
@@ -41,6 +40,7 @@ def create_heights_table(db, schema, table):
     """).format(schema=schema_q, table=table_q)
     try:
         db.sendQuery(query)
+        print("Created heights table")
         return(True)
     except:
         return(False)
@@ -51,8 +51,10 @@ def csv2db(db, cfg, args_in, out_paths):
     
     Note
     ----
-    Only for 3dfier's CSV-BUILDINGS-MULTIPLE output. Only works when the AHN3
-    and BAG tiles are the same (same size and identifier)
+    Only for 3dfier's CSV-BUILDINGS-MULTIPLE output. 
+    Only works when the AHN3 and BAG tiles are the same (same size and identifier). 
+    Only Linux.
+    Alter the CSV files by adding the ahn_file_date field and values.
     
     Parameters
     ----------
@@ -70,6 +72,8 @@ def csv2db(db, cfg, args_in, out_paths):
     
     schema_out_q = sql.Identifier(cfg['out_schema'])
     table_out_q = sql.Identifier(cfg['out_table'])
+    
+    table_idx = sql.Identifier(cfg['out_schema'] + "_id_idx")
     
     create_heights_table(db, cfg['out_schema'],
                                 cfg['out_table'])
@@ -91,31 +95,42 @@ def csv2db(db, cfg, args_in, out_paths):
                                            tile=tile_q)
                 cur.execute(query)
                 resultset = cur.fetchall()
-                ahn_file_date = resultset[0][0]
+                # the AHN3 file creation date that is stored in the tile index
+                ahn_file_date = resultset[0][0].isoformat()
+                
+                # Need to do some linux text-fu so that the whole csv file can
+                # be imported with COPY instead of row-wise edit and import
+                # in python (suuuper slow)
+                # Watch out for trailing commas from the CSV (until #58 is fixed in 3dfier)
+                command = "gawk -i inplace -F',' \
+                'BEGIN { OFS = \",\" } {$16=\"%s\"; print}' %s" % (ahn_file_date, p)
+                run(command, shell=True)
+                command = "sed -i '1s/.*/id,ground-0.00,ground-0.10,ground-0.20,\
+                ground-0.30,ground-0.40,ground-0.50,roof-0.00,roof-0.10,\
+                roof-0.25,roof-0.50,roof-0.75,roof-0.90,roof-0.95,roof-0.99,\
+                ahn_file_date/' %s" % p
+                run(command, shell=True)
                 
                 with open(p, "r") as f_in:
                     next(f_in) # skip header
-                    for row in f_in:
-                        f = StringIO(row + ahn_file_date.isoformat()) # trailing commas from the CSV (until #58 is fixed in 3dfier)
-                        cur.copy_from(f, tbl, sep=',')
+                    cur.copy_from(f_in, tbl, sep=',')
 
                 # delete the CSV
                 if args_in['rm']:
                     command = "rm" + p
-                    call(command, shell=True)
+                    run(command, shell=True)
                 else:
                     pass
     db.sendQuery(
-        sql.SQL("""CREATE INDEX IF NOT EXISTS {table}_id_idx
-                ON {schema}.{table} (id);
-                """.format(schema=schema_out_q,
-                           table=table_out_q)
-                )
+        sql.SQL("""CREATE INDEX {table}
+                ON {schema_q}.{table_q} (id);
+                """).format(schema_q=schema_out_q,
+                            table_q=table_out_q,
+                            table=table_idx)
     )
     db.sendQuery(
         sql.SQL("""COMMENT ON TABLE {schema}.{table} IS
                 'Building heights generated with 3dfier.';
-                """.format(schema=schema_out_q,
+                """).format(schema=schema_out_q,
                            table=table_out_q)
-                )
     )